git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/wxQT@65971 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ demos/life/life.cpp
@@ -354,10 +354,10 @@ void LifeFrame::UpdateInfoText()
 {
     wxString msg;
 
-    msg.Printf(_(" Generation: %u (T: %u ms),  Population: %u "),
+    msg.Printf(_(" Generation: %lu (T: %lu ms),  Population: %lu "),
                m_tics,
                m_topspeed? 0 : m_interval,
-               m_life->GetNumCells());
+               static_cast<unsigned long>(m_life->GetNumCells()));
     m_text->SetLabel(msg);
 }
